refactor: 아이템 타입 경로 수정
