+++ src/features/admin/model/data/index.ts
@@ -1,2 +1,1 @@
 export * from './admin-item-data';
-export * from './item-type';
