+++ src/features/admin/ui/form/AddItemForm.tsx
@@ -4,10 +4,10 @@ import { useNavigate } from 'react-router-dom';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { toast } from 'sonner';
 
-import { Button, Checkbox, Form, Input } from '@/shared';
+import { Button, Checkbox, Form, ITEM_TYPE, Input } from '@/shared';
 
 import { ItemField } from '../../components';
-import { AddItem, AddItemSchema, ITEM_TYPE } from '../../model';
+import { AddItem, AddItemSchema } from '../../model';
 
 export const AddItemForm = () => {
   //TODO: isPending 으로 변경
